requirements.txt
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.1
requests==2.32.3
ecdsa==0.19.0
eth_utils==4.1.1